fix: temporarily keep previous fee model in broadcaster
